requirements.txt
requests
httpx
hypha-rpc
python-dotenv
pyyaml
tqdm
boto3
shareloc-utils[potree]
pyyaml